model_server/requirements.txt
python-dotenv==1.0.0
transformers==4.32.1
datasets==2.14.4
accelerate==0.22.0
aio-pika==9.2.2